// apps/web/src/api/model/apiHydratedApiRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */
import type { ApiHydratedApiRequestQuery } from './apiHydratedApiRequestQuery';
import type { ApiHydratedApiRequestRequestBody } from './apiHydratedApiRequestRequestBody';
import type { ApiHydratedApiRequestStatus } from './apiHydratedApiRequestStatus';
import type { APIRequestAttempt } from './aPIRequestAttempt';
import type { AuditSource } from './auditSource';

export interface ApiHydratedApiRequest {
  /** Business ID */
  business: string;
  /** Secret key ID */
  secretKey?: string;
  /** Initiator user ID */
  initiator?: string;
  /** Request status */
  status: ApiHydratedApiRequestStatus;
  /** Request path */
  path: string;
  /** HTTP method */
  method: string;
  /** Query parameters */
  query?: ApiHydratedApiRequestQuery;
  /** Request body */
  requestBody?: ApiHydratedApiRequestRequestBody;
  /** Request attempts */
  attempts?: APIRequestAttempt[];
  /** Related event IDs */
  events?: string[];
  /** API version */
  apiVersion: string;
  /** Request source info */
  source?: AuditSource;
  /** Proxy source info */
  proxySource?: AuditSource;
  /** Document id */
  id: string;
  /** Document created at */
  createdAt: string;
  /** Document updated at */
  updatedAt: string;
}